+++ flutter/test/src/features/auth/data/auth_repository_impl_test.mocks.dart
@@ -68,17 +68,15 @@ class _FakeMacOsOptions_6 extends _i1.SmartFake implements _i3.MacOsOptions {
 /// See the documentation for Mockito's code generation for more information.
 class MockAuthRemoteDataSource extends _i1.Mock
     implements _i4.AuthRemoteDataSource {
+  MockAuthRemoteDataSource() {
+    _i1.throwOnMissingStub(this);
+  }
+
   @override
   _i5.Future<_i2.AuthResponse> login(String? email, String? password) =>
       (super.noSuchMethod(
             Invocation.method(#login, [email, password]),
             returnValue: _i5.Future<_i2.AuthResponse>.value(
-              _FakeAuthResponse_0(
-                this,
-                Invocation.method(#login, [email, password]),
-              ),
-            ),
-            returnValueForMissingStub: _i5.Future<_i2.AuthResponse>.value(
               _FakeAuthResponse_0(
                 this,
                 Invocation.method(#login, [email, password]),
@@ -110,16 +108,6 @@ class MockAuthRemoteDataSource extends _i1.Mock
                 ),
               ),
             ),
-            returnValueForMissingStub: _i5.Future<_i2.AuthResponse>.value(
-              _FakeAuthResponse_0(
-                this,
-                Invocation.method(
-                  #register,
-                  [email, password, name],
-                  {#invitationToken: invitationToken},
-                ),
-              ),
-            ),
           )
           as _i5.Future<_i2.AuthResponse>);
 
@@ -128,12 +116,6 @@ class MockAuthRemoteDataSource extends _i1.Mock
       (super.noSuchMethod(
             Invocation.method(#refresh, [refreshToken]),
             returnValue: _i5.Future<_i2.AuthResponse>.value(
-              _FakeAuthResponse_0(
-                this,
-                Invocation.method(#refresh, [refreshToken]),
-              ),
-            ),
-            returnValueForMissingStub: _i5.Future<_i2.AuthResponse>.value(
               _FakeAuthResponse_0(
                 this,
                 Invocation.method(#refresh, [refreshToken]),
@@ -157,15 +139,15 @@ class MockAuthRemoteDataSource extends _i1.Mock
 /// See the documentation for Mockito's code generation for more information.
 class MockFlutterSecureStorage extends _i1.Mock
     implements _i3.FlutterSecureStorage {
+  MockFlutterSecureStorage() {
+    _i1.throwOnMissingStub(this);
+  }
+
   @override
   _i3.IOSOptions get iOptions =>
       (super.noSuchMethod(
             Invocation.getter(#iOptions),
             returnValue: _FakeIOSOptions_1(this, Invocation.getter(#iOptions)),
-            returnValueForMissingStub: _FakeIOSOptions_1(
-              this,
-              Invocation.getter(#iOptions),
-            ),
           )
           as _i3.IOSOptions);
 
@@ -174,10 +156,6 @@ class MockFlutterSecureStorage extends _i1.Mock
       (super.noSuchMethod(
             Invocation.getter(#aOptions),
             returnValue: _FakeAndroidOptions_2(
-              this,
-              Invocation.getter(#aOptions),
-            ),
-            returnValueForMissingStub: _FakeAndroidOptions_2(
               this,
               Invocation.getter(#aOptions),
             ),
@@ -192,10 +170,6 @@ class MockFlutterSecureStorage extends _i1.Mock
               this,
               Invocation.getter(#lOptions),
             ),
-            returnValueForMissingStub: _FakeLinuxOptions_3(
-              this,
-              Invocation.getter(#lOptions),
-            ),
           )
           as _i3.LinuxOptions);
 
@@ -204,10 +178,6 @@ class MockFlutterSecureStorage extends _i1.Mock
       (super.noSuchMethod(
             Invocation.getter(#wOptions),
             returnValue: _FakeWindowsOptions_4(
-              this,
-              Invocation.getter(#wOptions),
-            ),
-            returnValueForMissingStub: _FakeWindowsOptions_4(
               this,
               Invocation.getter(#wOptions),
             ),
@@ -222,10 +192,6 @@ class MockFlutterSecureStorage extends _i1.Mock
               this,
               Invocation.getter(#webOptions),
             ),
-            returnValueForMissingStub: _FakeWebOptions_5(
-              this,
-              Invocation.getter(#webOptions),
-            ),
           )
           as _i3.WebOptions);
 
@@ -234,10 +200,6 @@ class MockFlutterSecureStorage extends _i1.Mock
       (super.noSuchMethod(
             Invocation.getter(#mOptions),
             returnValue: _FakeMacOsOptions_6(
-              this,
-              Invocation.getter(#mOptions),
-            ),
-            returnValueForMissingStub: _FakeMacOsOptions_6(
               this,
               Invocation.getter(#mOptions),
             ),
@@ -326,7 +288,6 @@ class MockFlutterSecureStorage extends _i1.Mock
               #wOptions: wOptions,
             }),
             returnValue: _i5.Future<String?>.value(),
-            returnValueForMissingStub: _i5.Future<String?>.value(),
           )
           as _i5.Future<String?>);
 
@@ -351,7 +312,6 @@ class MockFlutterSecureStorage extends _i1.Mock
               #wOptions: wOptions,
             }),
             returnValue: _i5.Future<bool>.value(false),
-            returnValueForMissingStub: _i5.Future<bool>.value(false),
           )
           as _i5.Future<bool>);
 
@@ -401,9 +361,6 @@ class MockFlutterSecureStorage extends _i1.Mock
             returnValue: _i5.Future<Map<String, String>>.value(
               <String, String>{},
             ),
-            returnValueForMissingStub: _i5.Future<Map<String, String>>.value(
-              <String, String>{},
-            ),
           )
           as _i5.Future<Map<String, String>>);
 
@@ -435,7 +392,6 @@ class MockFlutterSecureStorage extends _i1.Mock
       (super.noSuchMethod(
             Invocation.method(#isCupertinoProtectedDataAvailable, []),
             returnValue: _i5.Future<bool?>.value(),
-            returnValueForMissingStub: _i5.Future<bool?>.value(),
           )
           as _i5.Future<bool?>);
 }